Make cost functions into classes
