+++ .gitignore
@@ -19,3 +19,5 @@ src/**/*.js
 src/**/*.js.map
 tests/**/*.js
 tests/**/*.js.map
+
+*.swp
